:sparkles: Refs #547 - Get working with all registration backends
