+++ src/openforms/registrations/contrib/demo/__init__.py
@@ -1,5 +1,1 @@
 default_app_config = "openforms.registrations.contrib.demo.apps.DemoConfig"
-
-# import the converters to register them
-from . import converters  # noqa
-from .config import to_jsonschema  # noqa
