+++ src/openforms/registrations/contrib/email/config.py
@@ -2,8 +2,10 @@ from django.utils.translation import gettext_lazy as _
 
 from rest_framework import serializers
 
+from openforms.utils.jsonschema.mixins import JsonSchemaSerializerMixin
 
-class EmailOptionsSerializer(serializers.Serializer):
+
+class EmailOptionsSerializer(JsonSchemaSerializerMixin, serializers.Serializer):
     to_emails = serializers.ListField(
         child=serializers.EmailField(),
         label=_("The email addresses to which the submission details will be sent"),
