+++ src/openforms/registrations/contrib/zgw_apis/plugin.py
@@ -22,10 +22,11 @@ from openforms.registrations.contrib.zgw_apis.service import (
 from openforms.registrations.registry import register
 from openforms.submissions.mapping import FieldConf, apply_data_mapping
 from openforms.submissions.models import Submission, SubmissionReport
+from openforms.utils.jsonschema.mixins import JsonSchemaSerializerMixin
 from openforms.utils.validators import validate_rsin
 
 
-class ZaakOptionsSerializer(serializers.Serializer):
+class ZaakOptionsSerializer(JsonSchemaSerializerMixin, serializers.Serializer):
     zaaktype = serializers.URLField(
         required=False, help_text=_("URL of the ZAAKTYPE in the Catalogi API")
     )
