+++ src/openforms/utils/jsonschema/__init__.py
@@ -1,0 +1,3 @@
+# import the converters to register them
+from . import converters  # noqa
+from .convert import to_jsonschema  # noqa
